% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{blocking_time_series_split}
\alias{blocking_time_series_split}
\title{Time series cross-validation with no overlapping data}
\usage{
blocking_time_series_split(n, k, test_size)
}
\arguments{
\item{n}{length of time series}

\item{k}{number of folds}

\item{test_size}{fraction of samples in each fold to use as testing set.
Must be a number between 0 and 1 (exclusive).}
}
\value{
a generator that returns indices for fold 1,2,...,k
}
\description{
Time series cross-validation with no overlapping data
}
