% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxplot.cvsamples}
\alias{boxplot.cvsamples}
\title{Boxplots of cross-validation performances}
\usage{
\method{boxplot}{cvsamples}(x, ...)
}
\arguments{
\item{x}{a list containing models cross-validation performances,
using \code{create_samples} (from this package)}

\item{...}{additional parameters to be passed to \code{boxplot}}
}
\description{
Boxplots of cross-validation performances
}
\examples{

\dontrun{
print("see vignettes")
}


}
