% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_samples.R
\name{create_samples}
\alias{create_samples}
\title{Create a data structure of cross-validation results}
\usage{
create_samples(..., model_names)
}
\arguments{
\item{...}{list of cross-validation results for multiple models}

\item{model_names}{model names}
}
\value{
a list of results to be used in \code{\link{plot}}
}
\description{
Create a data structure of cross-validation results
}
\examples{

\dontrun{
print("see vignettes")
}

}
