% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\docType{class}
\name{CrossValidation-class}
\alias{CrossValidation-class}
\title{Gives groups of indices for cross validation}
\description{
Gives groups of indices for cross validation
}
\section{Slots}{

\describe{
\item{\code{k}}{number of folds}

\item{\code{validation}}{whether to have a validation set}

\item{\code{test}}{whether to have a test set}

\item{\code{data}}{contains the sample groupings (not intended for user)}
}}

