% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation-package.R
\docType{package}
\name{CrossValidation-package}
\alias{CrossValidation}
\alias{CrossValidation-package}
\title{CrossValidation: Generates Indices For Cross Validation}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
