% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{cv_split}
\alias{cv_split}
\title{Regular cross-validation split}
\usage{
cv_split(n, k)
}
\arguments{
\item{n}{number of items}

\item{k}{number of folds}
}
\value{
generator that returns the indices for fold 1,2,...,k
}
\description{
Regular cross-validation split
}
