% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{day_forward_chaining_split}
\alias{day_forward_chaining_split}
\title{Time series cross-validation with validation set}
\usage{
day_forward_chaining_split(n, k, whole_history = TRUE)
}
\arguments{
\item{n}{length of time series}

\item{k}{number of folds}

\item{whole_history}{whether to include all of the past at each fold}
}
\value{
generator that returns indices for fold 1,2,...,k
}
\description{
Time series cross-validation with validation set
}
