% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{get_fold}
\alias{get_fold}
\alias{get_fold,CrossValidation,integer-method}
\alias{get_fold,TimeSeriesCrossValidation,integer-method}
\title{Returns the ith fold of cross-validation}
\usage{
get_fold(cv, i)

\S4method{get_fold}{CrossValidation,integer}(cv, i)

\S4method{get_fold}{TimeSeriesCrossValidation,integer}(cv, i)
}
\arguments{
\item{cv}{CrossValidation-derived object}

\item{i}{which fold}
}
\value{
IndexTuple object containing the indices
   for train, validation, test sets
}
\description{
Returns the ith fold of cross-validation
}
