% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{get_folds}
\alias{get_folds}
\alias{get_folds,CrossValidation,integer-method}
\alias{get_folds,TimeSeriesCrossValidation,integer-method}
\title{Gets the folds needed for the train, validation, and test sets}
\usage{
get_folds(cv, i)

\S4method{get_folds}{CrossValidation,integer}(cv, i)

\S4method{get_folds}{TimeSeriesCrossValidation,integer}(cv, i)
}
\arguments{
\item{cv}{CrossValidation-derived object}

\item{i}{which fold}
}
\value{
list of folds to use in train, validation, and test sets
}
\description{
(Not intended for end user)
}
