% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\docType{class}
\name{IndexTuple-class}
\alias{IndexTuple-class}
\title{Represents a set of indices}
\description{
Represents a set of indices
}
\section{Slots}{

\describe{
\item{\code{train}}{the indices of items in the training set}

\item{\code{validation}}{the indices of items in the validation set}

\item{\code{test}}{the indices of items in the testing set}
}}

