% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{k_fold_cross_validation}
\alias{k_fold_cross_validation}
\title{Returns new k-fold cross-validation on n objects}
\usage{
k_fold_cross_validation(n, k, validation = FALSE, test = TRUE)
}
\arguments{
\item{n}{number of items}

\item{k}{number of folds}

\item{validation}{whether to have a validation set for each fold}

\item{test}{whether to have a test set for each fold}
}
\value{
KFoldCrossValidation object
}
\description{
Returns new k-fold cross-validation on n objects
}
