% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{k_fold_groups}
\alias{k_fold_groups}
\title{Generates k groups of roughly equal size}
\usage{
k_fold_groups(n, k)
}
\arguments{
\item{n}{number of items to be grouped}

\item{k}{number of groups}
}
\value{
vector of group membership of n items
}
\description{
Generates k groups of roughly equal size
}
