% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{plot_splits}
\alias{plot_splits}
\title{Plot samples in cross-validation scheme}
\usage{
plot_splits(splits)
}
\arguments{
\item{splits}{generator returned by other functions in this package}
}
\description{
Plot samples in cross-validation scheme
}
