% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{predict_lm}
\alias{predict_lm}
\title{Linear model prediction}
\usage{
predict_lm(fit_obj, newx)
}
\arguments{
\item{fit_obj}{object adjusted by \code{crossvalidation::fit_lm}}

\item{newx}{unseen data}
}
\value{
a vector or a matrix
}
\description{
Linear model prediction
}
\examples{

NULL

}
