% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folds.R
\name{split_ts}
\alias{split_ts}
\title{Split a time series}
\usage{
split_ts(y, p = 0.8, return_indices = FALSE)
}
\arguments{
\item{y}{univariate or multivariate time series}

\item{p}{proportion of data in training set}

\item{return_indices}{return indices instead of time series?}
}
\description{
Split a time series
}
