% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{stratified_cv_split}
\alias{stratified_cv_split}
\title{Stratified cross-validation}
\usage{
stratified_cv_split(groups, k)
}
\arguments{
\item{groups}{factor of group membership of items}

\item{k}{number of groups}
}
\value{
generator that returns the indices for fold 1,2,...,k
}
\description{
Stratified cross-validation
}
