% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{stratified_k_fold_cross_validation}
\alias{stratified_k_fold_cross_validation}
\title{Stratified k-fold cross-validation}
\usage{
stratified_k_fold_cross_validation(groups, k, validation = FALSE, test = TRUE)
}
\arguments{
\item{groups}{factor of the groups each sample belongs to}

\item{k}{number of folds of cross-validation}

\item{validation}{whether to have a validation set for each fold}

\item{test}{whether to have a test set for each fold}
}
\value{
StratifiedKFoldCrossValidation object
}
\description{
Stratified cross-validation yields sets of samples
in roughly the same proportions as in the original dataset.
}
