% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\docType{class}
\name{StratifiedKFoldCrossValidation-class}
\alias{StratifiedKFoldCrossValidation-class}
\title{k-fold cross validation that respects group frequencies}
\description{
k-fold cross validation that respects group frequencies
}
