% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{time_series_cross_validation}
\alias{time_series_cross_validation}
\title{Time series cross-validation}
\usage{
time_series_cross_validation(
  n,
  k,
  validation = FALSE,
  test = TRUE,
  all_preceding_periods = TRUE
)
}
\arguments{
\item{n}{length of time series}

\item{k}{number of folds of cross-validation}

\item{validation}{whether validation set is needed}

\item{test}{whether test set is needed}

\item{all_preceding_periods}{whether to include all history up to
the current training set.}
}
\value{
TimeSeriesCrossValidation object
}
\description{
Time series cross-validation
}
