% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{time_series_split}
\alias{time_series_split}
\title{Regular time series cross-validation}
\usage{
time_series_split(n, k, whole_history = TRUE)
}
\arguments{
\item{n}{length of time series}

\item{k}{number of folds}

\item{whole_history}{whether to include all past data in the training set}
}
\value{
generator that returns indices for fold 1,2,...,k
}
\description{
Regular time series cross-validation
}
