% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\docType{class}
\name{TimeSeriesCrossValidation-class}
\alias{TimeSeriesCrossValidation-class}
\title{CrossValidation that respects time series order}
\description{
CrossValidation that respects time series order
}
\section{Slots}{

\describe{
\item{\code{all_preceding_periods}}{whether to include all preceding periods
when constructing training set}
}}

