% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval_ts.R
\name{crossval_ts}
\alias{crossval_ts}
\title{Generic cross-validation function for time series}
\usage{
crossval_ts(
  y,
  x = NULL,
  fit_func = crossval::fit_lm,
  predict_func = crossval::predict_lm,
  fcast_func = NULL,
  fit_params = NULL,
  initial_window = 5,
  horizon = 3,
  fixed_window = TRUE,
  seed = 123,
  eval_metric = NULL,
  cl = NULL,
  errorhandling = c("stop", "remove", "pass"),
  packages = c("stats", "Rcpp"),
  verbose = FALSE,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{y}{response time series; a vector}

\item{x}{input covariates' matrix (optional)}

\item{fit_func}{a function for fitting the model}

\item{predict_func}{a function for predicting values from the model}

\item{fcast_func}{time series forecasting function}

\item{fit_params}{a list; additional (model-specific) parameters to be passed
to \code{fit_func}}

\item{initial_window}{an integer; the initial number of consecutive values in each training set sample}

\item{horizon}{an integer; the number of consecutive values in test set sample}

\item{fixed_window}{a boolean; if FALSE, all training samples start at 1}

\item{seed}{random seed for reproducibility of results}

\item{eval_metric}{a function measuring the test errors; if not provided: RMSE for regression and
accuracy for classification}

\item{cl}{an integer; the number of clusters for parallel execution}

\item{errorhandling}{specifies how a task evalution error should be handled.
If value is "stop", then execution will be stopped if an error occurs. If value
is "remove", the result for that task will not be returned. If value is "pass",
then the error object generated by task evaluation will be included with the
rest of the results. The default value is "stop".}

\item{packages}{character vector of packages that the tasks depend on}

\item{verbose}{logical flag enabling verbose messages. This can be very useful for
troubleshooting.}

\item{show_progress}{show evolution of the algorithm}

\item{...}{additional parameters}
}
\value{

}
\description{
Generic cross-validation for univariate time series
}
\examples{


require(forecast)
data("AirPassengers")

# Example 1 -----

res <- crossval_ts(y=AirPassengers, initial_window = 10, fcast_func = thetaf)
print(colMeans(res))


# Example 2 -----

fcast_func <- function (y, h, ...)
{
      forecast::forecast(forecast::auto.arima(y, ...),
      h=h, ...)
}

res <- crossval_ts(y=AirPassengers, initial_window = 10, fcast_func = fcast_func)
print(colMeans(res))


# Example 3 -----

fcast_func <- function (y, h, ...)
{
      forecast::forecast(forecast::ets(y, ...),
      h=h, ...)
}

res <- crossval_ts(y=AirPassengers, initial_window = 10, fcast_func = fcast_func)
print(colMeans(res))


# Example 4 -----

xreg <- cbind(1, 1:length(AirPassengers))
res <- crossval_ts(y=AirPassengers, x=xreg, fit_func = crossval::fit_lm,
predict_func = crossval::predict_lm,
initial_window = 10,
horizon = 3,
fixed_window = TRUE)
colMeans(res)

}
